Validation Fix

+++ Frontend/src/stores/CategoryStore.ts
@@ -4,6 +4,7 @@ import { APIResponse } from '../models/APIResult';
 import CategoryService from '../services/CategoryService';
 import { CategoryCreateDTO, CategoryGetDTO } from '../models/CategoryDtos';
 import { useAuthStore } from './AuthStore';
+import Login from '../views/Auth/Login.vue';
 
 const toast = useToast();
 export const useCategoryStore = defineStore('category', {
@@ -13,33 +14,52 @@ export const useCategoryStore = defineStore('category', {
   }),
   actions:{
     async fetchCategories() {
-      try {
-        const response: APIResponse<any> = await CategoryService.getAllCategories();
-        this.categories = response.Result;
-      } catch (error) {
-        toast.error('Error fetching categories');
+      const authStore=useAuthStore()
+      if (authStore.isLoggedIn) {
+        try {
+          const response: APIResponse<any> = await CategoryService.getAllCategories();
+          this.categories = response.Result;
+        } catch (error) {
+          toast.error('Error fetching categories');
+        }        
       }
+      else{
+        toast.error('You must be logged in to view categories');
+      }
+
     },
     async addCategory() {
       const authStore = useAuthStore();
-      try {
-        await CategoryService.createCategory(this.currentCategory, authStore.token);
-        this.fetchCategories();
-        toast.success('Category added successfully');
-      } catch (error) {
-        toast.error('Error adding category');
+      if (authStore.isLoggedIn &&authStore.role==='admin') {
+        try {
+          await CategoryService.createCategory(this.currentCategory, authStore.token);
+          this.fetchCategories();
+          toast.success('Category added successfully');
+        } catch (error) {
+          toast.error('Error adding category');
+        }
+      }
+      else{
+        toast.error('You must be logged in to add a category');
       }
     },
+
+     
     async deleteBurger(id: number) {
       const authStore = useAuthStore();
-      try {
-        await CategoryService.deleteCategory(id, authStore.token);
-        this.categories = this.categories.filter(u => u.Id !== id);
-        toast.success('Burger deleted successfully');
-      } catch (error) {
-        toast.error('Error deleting burger');
+      if (authStore.isLoggedIn && authStore.role==='admin') {
+        try {
+          await CategoryService.deleteCategory(id, authStore.token);
+          this.categories = this.categories.filter(u => u.Id !== id);
+          toast.success('Burger deleted successfully');
+        } catch (error) {
+          toast.error('Error deleting burger');
+        }
       }
-    },
+      else{
+        toast.error('You must be logged in to delete a burger');
+      }
+    }
   }
 })
 export default useCategoryStore;
